device-class: Introduce pa_device_class_t
My plan is to use the device class in node names. There is already the
PA_PROP_DEVICE_FORM_FACTOR property, which is almost the same thing as
the device class, but not exactly the same. A conversion function from
the device class to the form factor is provided.

I wouldn't be surprised if further uses for the device class
enumeration were found in the future.