Create private null sinks for sink input nodes that aren't routed anywhere
Currently this only works for the initial routing. If a sink input
node is first routed to some real output, and then the routing is
changed so that the sink input node is not any more connected
anywhere, the private sink isn't used even though it should be.
I don't actually know what happens in the scenario where the routing
system decides to leave a sink input without connections (other than
in the initial routing case)...

This changes the default policy: previously, if the sink input routing
was not decided by the sink input creator or policy modules, the sink
input would get routed to the default sink as a fallback. Now the
fallback is a null sink. I intend to change this so that if there are
no routing modules loaded, the core will implement a default router
that routes sink inputs to the default sink.