tunnel-new: Read parameters from tunnel manager when available
This prepares for loading tunnel sink and source module instances from
the tunnel manager.

If the tunnel manager is available, and the tunnel manager has
sample spec and channel map information for the device, then let's use
that information when creating the sink or source so that the tunnel
will automatically match the parameters of the remote device. The
remote device's proplist is copied to the tunnel too.

Change-Id: I2755d6fc984d9014bc039924dd1d51ea5689c22c