logind: Move the code that interfaces with logind to liblogind.so
This doesn't change the behaviour in any important way. The purpose of
this move is to provide access to logind from multiple modules without
needing to use the logind API from more than one place. Currently
there's only one module using liblogind.so, but module-tunnel-manager
will need seat tracking, so that will be another module that needs to
access logind.

Change-Id: I352193fef0d0b0601ea509af501af395786c6cf9