tunnel-manager: Make the tunnel enabling policy configurable
There's no real configurability yet, because only one condition is
supported. Support for more conditions will be added later.

The syntax of the configuration option pretends to conform to some
programming language, but there's no real parser for the language.
A proper parser may be implemented later, but for now the option
values are parsed simply by comparing to a list of predefined string
constants.

Change-Id: I57bbb0cd2b91691415be78f9732aef6b3bebde56