pulsecore: added reliable PID tracking for (local) clients.
Automatically retrieve and store client credentials if those
are available and enabled. Credentials will be made available
for clients that connect using a socket/protocol family that
is capable of passing credentials (UNIX domain sockets in
practice).

Currently credentials will not be be available until the first
protocol message is received. If it turns out that credentials
are needed immediately after connection establishment, it is
easy to extend pstream to explictly query the credentials over
its I/O channel.

Although not used at the moment, since we store the full set of
retrieved credentials, we get uid and gid tracking in addition
to PID for free, as a side effect. Appropriate functions have
been added to query those as well.

Change-Id: I8ea306761e0b901a1bdf3ec9203920458162d670
Signed-off-by: Krisztian Litkey <kli@iki.fi>