pulsecore: added pa_client_getenv.
Added pa_client_getenv, the core function to retrieve white-
listed variables from a clients environment. The actual code
to retrieve variables is not in core. One implementation for
it is available in module-expose-environment.

Alternative implementations with more or different logic or
restrictions is easily possible by loading a different module
that hooks itself up to the same minimalistic core infra for
this.

Change-Id: I60e784268e8bcbde5d2f9a3f6aefd48b2254badb
Signed-off-by: Krisztian Litkey <kli@iki.fi>